Convert lua table from stack to JSON
Can now read a table from the Lua stack and return a json::Object from it.
Lua "Arrays" are read into json::Objects too. Any json::Object has a list of it's children si it's basically an array.

the function to be used is:
json::Node* fromLuaStack(lua_State* L, int i);

The table should be on the top of the stack.